New Cable management
This is a photo of the new wire/cable management. They are neatly tucked into and around the RPi, instead of being connected directly into the terminals from the motors with the wires sticking out all the sides. This is an edited version of the earlier commit that Brandon has now deleted as the file name of the photo displayed the wrong week, and I thought this would be confusing.