Updated Robot Hardware
This is the robot's current state at the beginning of the week four lab. I printed pieces to attach the camera to, so that it had a stand, which Brandon attached.
It was wired last week, but into the wrong terminals. We fixed that but not with good cable management as we were just trying to get it working. Brandon also tidied up the cables in spare time.