To reverse and turn when at a dead-end in the maze
This code will hopefully allow the AV to reverse once detecting a dead end within the maze, and then will turn left or right depending on an open space.
The reversing has been coded, and needs to be tested. Still working on the turning when sensors pick up that there is enough room for the bot to turn.